Further cleanup addressing more copied code
