+++ src/main/java/hudson/security/AuthorizationMatrixProperty.java
@@ -46,10 +46,11 @@ import jenkins.model.Jenkins;
 import net.sf.json.JSONObject;
 import org.acegisecurity.acls.sid.PrincipalSid;
 import org.acegisecurity.acls.sid.Sid;
-import org.jenkinsci.plugins.matrixauth.AbstractMatrixPropertyConverter;
-import org.jenkinsci.plugins.matrixauth.AuthorizationMatrixPropertyDescriptor;
+import org.jenkinsci.plugins.matrixauth.AbstractAuthorizationPropertyConverter;
+import org.jenkinsci.plugins.matrixauth.AuthorizationPropertyDescriptor;
 import org.jenkinsci.plugins.matrixauth.AuthorizationProperty;
 import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.DoNotUse;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
 import org.jenkinsci.plugins.matrixauth.inheritance.InheritParentStrategy;
 import org.jenkinsci.plugins.matrixauth.inheritance.InheritanceStrategy;
@@ -65,67 +66,66 @@ import org.kohsuke.stapler.StaplerRequest;
  */
 public class AuthorizationMatrixProperty extends JobProperty<Job<?, ?>> implements AuthorizationProperty {
 
-	private final transient SidACL acl = new AclImpl();
+    private final transient SidACL acl = new AclImpl();
 
-	/**
-	 * List up all permissions that are granted.
-	 * 
-	 * Strings are either the granted authority or the principal, which is not
-	 * distinguished.
-	 */
-	private final Map<Permission, Set<String>> grantedPermissions = new HashMap<>();
+    /**
+     * List up all permissions that are granted.
+     * <p>
+     * Strings are either the granted authority or the principal, which is not
+     * distinguished.
+     */
+    private final Map<Permission, Set<String>> grantedPermissions = new HashMap<>();
 
-	private final Set<String> sids = new HashSet<>();
+    private final Set<String> sids = new HashSet<>();
 
-	/**
-	 * @deprecated unused, use {@link #setInheritanceStrategy(InheritanceStrategy)} instead.
-	 */
-	@Deprecated
+    /**
+     * @deprecated unused, use {@link #setInheritanceStrategy(InheritanceStrategy)} instead.
+     */
+    @Deprecated
     private transient Boolean blocksInheritance;
 
-	private InheritanceStrategy inheritanceStrategy = new InheritParentStrategy();
+    private InheritanceStrategy inheritanceStrategy = new InheritParentStrategy();
 
     private AuthorizationMatrixProperty() {
     }
 
     public AuthorizationMatrixProperty(Map<Permission, Set<String>> grantedPermissions) {
         // do a deep copy to be safe
-        for (Entry<Permission,Set<String>> e : grantedPermissions.entrySet())
-            this.grantedPermissions.put(e.getKey(),new HashSet<>(e.getValue()));
+        for (Entry<Permission, Set<String>> e : grantedPermissions.entrySet())
+            this.grantedPermissions.put(e.getKey(), new HashSet<>(e.getValue()));
     }
 
-	public Set<String> getGroups() {
-		return sids;
-	}
+    public Set<String> getGroups() {
+        return sids;
+    }
 
     /**
      * Returns all the (Permission,sid) pairs that are granted, in the multi-map form.
      *
-     * @return
-     *      read-only. never null.
+     * @return read-only. never null.
      */
-    public Map<Permission,Set<String>> getGrantedPermissions() {
+    public Map<Permission, Set<String>> getGrantedPermissions() {
         return Collections.unmodifiableMap(grantedPermissions);
     }
 
     /**
-	 * Adds to {@link #grantedPermissions}. Use of this method should be limited
-	 * during construction, as this object itself is considered immutable once
-	 * populated.
-	 */
-	public void add(Permission p, String sid) {
-		Set<String> set = grantedPermissions.get(p);
-		if (set == null)
-			grantedPermissions.put(p, set = new HashSet<>());
-		set.add(sid);
-		sids.add(sid);
-	}
+     * Adds to {@link #grantedPermissions}. Use of this method should be limited
+     * during construction, as this object itself is considered immutable once
+     * populated.
+     */
+    public void add(Permission p, String sid) {
+        Set<String> set = grantedPermissions.get(p);
+        if (set == null)
+            grantedPermissions.put(p, set = new HashSet<>());
+        set.add(sid);
+        sids.add(sid);
+    }
 
     @Extension
-    public static class DescriptorImpl extends JobPropertyDescriptor implements AuthorizationMatrixPropertyDescriptor<AuthorizationMatrixProperty> {
+    public static class DescriptorImpl extends JobPropertyDescriptor implements AuthorizationPropertyDescriptor<AuthorizationMatrixProperty> {
 
         @Override
-        public AuthorizationMatrixProperty createProperty() {
+        public AuthorizationMatrixProperty create() {
             return new AuthorizationMatrixProperty();
         }
 
@@ -133,97 +133,98 @@ public class AuthorizationMatrixProperty extends JobProperty<Job<?, ?>> implemen
         public PermissionScope getPermissionScope() {
             return PermissionScope.ITEM;
         }
-        
-		@Override
-		public JobProperty<?> newInstance(StaplerRequest req, JSONObject formData) throws FormException {
-            return createNewInstance(req, formData, true);
-		}
 
-		@Override
-		public boolean isApplicable(Class<? extends Job> jobType) {
+        @Override
+        public JobProperty<?> newInstance(StaplerRequest req, JSONObject formData) throws FormException {
+            return createNewInstance(req, formData, true);
+        }
+
+        @Override
+        public boolean isApplicable(Class<? extends Job> jobType) {
             return isApplicable();
-		}
+        }
 
         public FormValidation doCheckName(@AncestorInPath Job project, @QueryParameter String value) {
-            return GlobalMatrixAuthorizationStrategy.DESCRIPTOR.doCheckName_(value, project, Item.CONFIGURE);
+            return doCheckName_(value, project, Item.CONFIGURE);
         }
     }
 
-	private final class AclImpl extends SidACL {
-                @CheckForNull
-                @SuppressFBWarnings(value = "NP_BOOLEAN_RETURN_NULL", 
-                        justification = "As designed, implements a third state for the ternary logic")
-		protected Boolean hasPermission(Sid sid, Permission p) {
-			if (AuthorizationMatrixProperty.this.hasPermission(toString(sid),p,sid instanceof PrincipalSid)) {
-				return true;
+    private final class AclImpl extends SidACL {
+        @CheckForNull
+        @SuppressFBWarnings(value = "NP_BOOLEAN_RETURN_NULL",
+                justification = "As designed, implements a third state for the ternary logic")
+        protected Boolean hasPermission(Sid sid, Permission p) {
+            if (AuthorizationMatrixProperty.this.hasPermission(toString(sid), p, sid instanceof PrincipalSid)) {
+                return true;
+            }
+            return null;
+        }
+    }
+
+    public SidACL getACL() {
+        return acl;
+    }
+
+    public void setInheritanceStrategy(InheritanceStrategy inheritanceStrategy) {
+        this.inheritanceStrategy = inheritanceStrategy;
+    }
+
+    public InheritanceStrategy getInheritanceStrategy() {
+        return inheritanceStrategy;
+    }
+
+    /**
+     * Persist {@link ProjectMatrixAuthorizationStrategy} as a list of IDs that
+     * represent {@link ProjectMatrixAuthorizationStrategy#grantedPermissions}.
+     */
+    @Restricted(DoNotUse.class)
+    public static final class ConverterImpl extends AbstractAuthorizationPropertyConverter<AuthorizationMatrixProperty> {
+        public boolean canConvert(Class type) {
+            return type == AuthorizationMatrixProperty.class;
+        }
+
+        public AuthorizationMatrixProperty create() {
+            return new AuthorizationMatrixProperty();
+        }
+    }
+
+    /**
+     * Ensure that the user creating a job has Read and Configure permissions
+     */
+    @Extension
+    @Restricted(NoExternalUse.class)
+    public static class ItemListenerImpl extends ItemListener {
+        @Override
+        public void onCreated(Item item) {
+            AuthorizationStrategy authorizationStrategy = Jenkins.getInstance().getAuthorizationStrategy();
+            if (authorizationStrategy instanceof ProjectMatrixAuthorizationStrategy) {
+                ProjectMatrixAuthorizationStrategy strategy = (ProjectMatrixAuthorizationStrategy) authorizationStrategy;
+
+                if (item instanceof Job) {
+                    Job<?, ?> job = (Job<?, ?>) item;
+                    AuthorizationMatrixProperty prop = job.getProperty(AuthorizationMatrixProperty.class);
+                    if (prop == null) {
+                        prop = new AuthorizationMatrixProperty();
+                    }
+
+                    User current = User.current();
+                    String sid = current == null ? "anonymous" : current.getId();
+
+                    if (!strategy.getACL(job).hasPermission(Jenkins.getAuthentication(), Item.READ)) {
+                        prop.add(Item.READ, sid);
+                    }
+                    if (!strategy.getACL(job).hasPermission(Jenkins.getAuthentication(), Item.CONFIGURE)) {
+                        prop.add(Item.CONFIGURE, sid);
+                    }
+                    if (prop.getGrantedPermissions().size() > 0) {
+                        try {
+                            job.addProperty(prop);
+                        } catch (IOException ex) {
+                            // TODO LOGGER
                         }
-			return null;
-		}
-	}
-
-	public SidACL getACL() {
-		return acl;
-	}
-
-	public void setInheritanceStrategy(InheritanceStrategy inheritanceStrategy) {
-		this.inheritanceStrategy = inheritanceStrategy;
-	}
-
-	public InheritanceStrategy getInheritanceStrategy() {
-		return inheritanceStrategy;
-	}
-
-	/**
-	 * Persist {@link ProjectMatrixAuthorizationStrategy} as a list of IDs that
-	 * represent {@link ProjectMatrixAuthorizationStrategy#grantedPermissions}.
-	 */
-	public static final class ConverterImpl extends AbstractMatrixPropertyConverter {
-		public boolean canConvert(Class type) {
-			return type == AuthorizationMatrixProperty.class;
-		}
-
-		public AuthorizationProperty createSubject() {
-		    return new AuthorizationMatrixProperty();
+                    }
+                }
+            }
         }
     }
-
-	/**
-	 * Ensure that the user creating a job has Read and Configure permissions
-	 */
-	@Extension
-	@Restricted(NoExternalUse.class)
-	public static class ItemListenerImpl extends ItemListener {
-		@Override
-		public void onCreated(Item item) {
-			AuthorizationStrategy authorizationStrategy = Jenkins.getInstance().getAuthorizationStrategy();
-			if (authorizationStrategy instanceof ProjectMatrixAuthorizationStrategy) {
-				ProjectMatrixAuthorizationStrategy strategy = (ProjectMatrixAuthorizationStrategy) authorizationStrategy;
-
-				if (item instanceof Job) {
-					Job<?, ?> job = (Job<?, ?>) item;
-					AuthorizationMatrixProperty prop = job.getProperty(AuthorizationMatrixProperty.class);
-					if (prop == null) {
-						prop = new AuthorizationMatrixProperty();
-					}
-
-					User current = User.current();
-					String sid = current == null ? "anonymous" : current.getId();
-
-					if (!strategy.getACL(job).hasPermission(Jenkins.getAuthentication(), Item.READ)) {
-						prop.add(Item.READ, sid);
-					}
-					if (!strategy.getACL(job).hasPermission(Jenkins.getAuthentication(), Item.CONFIGURE)) {
-						prop.add(Item.CONFIGURE, sid);
-					}
-					if (prop.getGrantedPermissions().size() > 0) {
-						try {
-							job.addProperty(prop);
-						} catch (IOException ex) {
-							// TODO LOGGER
-						}
-					}
-				}
-			}
-		}
-	}
 }
