+++ src/main/java/org/jenkinsci/plugins/matrixauth/AuthorizationProperty.java
@@ -23,26 +23,11 @@
  */
 package org.jenkinsci.plugins.matrixauth;
 
-import hudson.security.GlobalMatrixAuthorizationStrategy;
-import hudson.security.Permission;
-import hudson.security.SecurityRealm;
-import jenkins.model.IdStrategy;
-import jenkins.model.Jenkins;
 import org.jenkinsci.plugins.matrixauth.inheritance.InheritGlobalStrategy;
 import org.jenkinsci.plugins.matrixauth.inheritance.InheritanceStrategy;
 import org.jenkinsci.plugins.matrixauth.inheritance.NonInheritingStrategy;
-import org.kohsuke.accmod.Restricted;
-import org.kohsuke.accmod.restrictions.NoExternalUse;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.TreeSet;
-
-public interface AuthorizationProperty {
-    void add(Permission permission, String sid);
-    Map<Permission, Set<String>> getGrantedPermissions();
+public interface AuthorizationProperty extends AuthorizationContainer {
 
     void setInheritanceStrategy(InheritanceStrategy inheritanceStrategy);
     InheritanceStrategy getInheritanceStrategy();
@@ -57,9 +42,8 @@ public interface AuthorizationProperty {
      *
      * Note that for items nested inside folders, this will change behavior significantly.
      *
-     * @param blocksInheritance
      * @since 2.0
-     * @deprecated
+     * @deprecated Use {@link InheritanceStrategy} instead.
      */
     @Deprecated
     default void setBlocksInheritance(boolean blocksInheritance) {
@@ -77,115 +61,12 @@ public interface AuthorizationProperty {
      * Since the introduction of inheritance strategies, returns {@code true}
      * if and only if the selected inheritance strategy is {@link NonInheritingStrategy}.
      *
-     * @return
      * @since 2.0
-     * @deprecated
+     * @deprecated Use {@link #getInheritanceStrategy()} instead.
      */
     @Deprecated
     default boolean isBlocksInheritance() {
         return getInheritanceStrategy() instanceof NonInheritingStrategy;
     }
 
-    /**
-     * Checks if the given SID has the given permission.
-     */
-    default boolean hasPermission(String sid, Permission p) {
-        final SecurityRealm securityRealm = Jenkins.getInstance().getSecurityRealm();
-        final IdStrategy groupIdStrategy = securityRealm.getGroupIdStrategy();
-        final IdStrategy userIdStrategy = securityRealm.getUserIdStrategy();
-        for (; p != null; p = p.impliedBy) {
-            if (!p.getEnabled()) {
-                continue;
-            }
-            Set<String> set = getGrantedPermissions().get(p);
-            if (set != null && set.contains(sid))
-                return true;
-            if (set != null) {
-                for (String s: set) {
-                    if (userIdStrategy.equals(s, sid) || groupIdStrategy.equals(s, sid)) {
-                        return true;
-                    }
-                }
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Checks if the given SID has the given permission.
-     */
-    default boolean hasPermission(String sid, Permission p, boolean principal) {
-        final SecurityRealm securityRealm = Jenkins.getInstance().getSecurityRealm();
-        final IdStrategy strategy = principal ? securityRealm.getUserIdStrategy() : securityRealm.getGroupIdStrategy();
-        for (; p != null; p = p.impliedBy) {
-            if (!p.getEnabled()) {
-                continue;
-            }
-            Set<String> set = getGrantedPermissions().get(p);
-            if (set == null) {
-                continue;
-            }
-            if (set.contains(sid)) {
-                return true;
-            }
-            for (String s : set) {
-                if (strategy.equals(s, sid)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Checks if the permission is explicitly given, instead of implied through {@link Permission#impliedBy}.
-     */
-    default boolean hasExplicitPermission(String sid, Permission p) {
-        if (sid == null) { // used for template row in UI
-            return false;
-        }
-        Set<String> set = getGrantedPermissions().get(p);
-        if (set != null && p.getEnabled()) {
-            if (set.contains(sid))
-                return true;
-            final SecurityRealm securityRealm = Jenkins.getInstance().getSecurityRealm();
-            final IdStrategy groupIdStrategy = securityRealm.getGroupIdStrategy();
-            final IdStrategy userIdStrategy = securityRealm.getUserIdStrategy();
-            for (String s: set) {
-                if (userIdStrategy.equals(s, sid) || groupIdStrategy.equals(s, sid)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Works like {@link #add(Permission, String)} but takes both parameters
-     * from a single string of the form <tt>PERMISSIONID:sid</tt>
-     */
-    @Restricted(NoExternalUse.class)
-    default void add(String shortForm) {
-        int idx = shortForm.indexOf(':');
-        Permission p = Permission.fromId(shortForm.substring(0, idx));
-        if (p==null)
-            throw new IllegalArgumentException("Failed to parse '"+shortForm+"' --- no such permission");
-        add(p, shortForm.substring(idx + 1));
-    }
-
-    /**
-     * Returns all SIDs configured in this matrix, minus "anonymous"
-     *
-     * @return Always non-null.
-     */
-    default List<String> getAllSIDs() {
-        Set<String> r = new TreeSet<>(new GlobalMatrixAuthorizationStrategy.IdStrategyComparator());
-        for (Set<String> set : getGrantedPermissions().values())
-            r.addAll(set);
-        r.remove("anonymous");
-
-        String[] data = r.toArray(new String[r.size()]);
-        Arrays.sort(data);
-        return Arrays.asList(data);
-    }
 }
